requirements.txt
torch==2.3.0
timm==1.0.3
numpy==1.26.4
huggingface_hub==0.23.1
einops==0.8.0
transformers==4.41.1
open_clip_torch==2.24.0